fix: reduce bundle size
