+++ src/app.tsx
@@ -1,34 +1,7 @@
-import { RouterProvider, createRouter } from "@tanstack/react-router";
-
-import ApiClientWrapper from "@common/api/api-client-wrapper";
-import { useSession } from "@common/api/use-session";
-
-import { routeTree } from "./routeTree.gen";
-
-const router = createRouter({
-	routeTree,
-	context: {
-		session: null,
-	},
-});
-
-declare module "@tanstack/react-router" {
-	interface Register {
-		router: typeof router;
-	}
-}
-
-const AppWithAuth = () => {
-	const session = useSession();
-	return <RouterProvider router={router} context={{ session }} />;
-};
+import { Router } from "./router";
 
 const App = () => {
-	return (
-		<ApiClientWrapper>
-			<AppWithAuth />
-		</ApiClientWrapper>
-	);
+	return <Router />;
 };
 
 export default App;
