+++ src/common/components/app-wrapper.tsx
@@ -1,7 +1,5 @@
 import { MantineProvider } from "@mantine/core";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { TanStackRouterDevtools } from "@tanstack/react-router-devtools";
 import { PropsWithChildren, useRef } from "react";
 
 import { getQueryClient } from "@common/api/query-client";
@@ -16,10 +14,8 @@ const AppWrapper = ({ children }: PropsWithChildren) => {
 			<MantineProvider theme={theme} defaultColorScheme="light">
 				<QueryClientProvider client={queryClient.current}>
 					{children}
-					<ReactQueryDevtools />
 				</QueryClientProvider>
 			</MantineProvider>
-			<TanStackRouterDevtools />
 		</>
 	);
 };
