+++ src/common/types/router.ts
@@ -1,0 +1,3 @@
+export interface RouterContext {
+	isAuthenticated: boolean;
+}
