+++ src/feature/submissions/create-beer-location-submission.dialog.tsx
@@ -1,95 +1,15 @@
-import { Modal, Title, Notification, Button } from "@mantine/core";
-import { IconBeer, IconCheck } from "@tabler/icons-react";
-import { useState } from "react";
+import ApiClientWrapper from "@common/api/api-client-wrapper";
 
-import { useDistricts } from "@common/api/queries";
-import { BeerLocationForm } from "@common/components/beer-location.form";
-import type { BeerLocationFormData } from "@common/types/beer-location-form-data";
-import type { NotificationType } from "@common/types/common";
+import BeerLocationSubmissionDialogContent from "./create-beer-location-submission.content";
 
-import { useCreateBeerLocationSubmission } from "./queries";
+interface Props {
+	onClose: () => void;
+}
 
-export const BeerLocationSubmissionDialog = () => {
-	const [open, setOpen] = useState(false);
-	const mutation = useCreateBeerLocationSubmission();
-	const { data: districts } = useDistricts();
-	const [showNotification, setShowNotification] = useState<
-		NotificationType | undefined
-	>(undefined);
+const BeerLocationSubmissionDialog = ({ onClose }: Props) => (
+	<ApiClientWrapper>
+		<BeerLocationSubmissionDialogContent onClose={onClose} />
+	</ApiClientWrapper>
+);
 
-	const onSubmit = async (data: BeerLocationFormData) => {
-		try {
-			await mutation.mutateAsync(data);
-			setShowNotification("success");
-			setOpen(false);
-			setTimeout(() => {
-				setShowNotification(undefined);
-			}, 2000);
-		} catch (e) {
-			console.error(e);
-			setShowNotification("error");
-			setTimeout(() => {
-				setShowNotification(undefined);
-			}, 3000);
-		}
-	};
-
-	return (
-		<>
-			<Button
-				leftSection={<IconBeer size={16} />}
-				onClick={() => {
-					setOpen(true);
-				}}
-			>
-				Föreslå ny plats
-			</Button>
-
-			{showNotification && (
-				<Notification
-					title={
-						showNotification === "success"
-							? "Förslag skickat!"
-							: "Något gick fel"
-					}
-					color={showNotification === "error" ? "red" : "green"}
-					styles={{
-						root: {
-							position: "fixed",
-							top: "2rem",
-							right: "2rem",
-							zIndex: 9999999,
-						},
-					}}
-					withBorder
-					withCloseButton
-					onClose={() => {
-						setShowNotification(undefined);
-					}}
-					icon={<IconCheck />}
-				>
-					{showNotification === "success"
-						? "Ditt förslag på ny plats har skickats för granskning"
-						: "Försök igen eller kontakta en katt om problemet kvarstår"}
-				</Notification>
-			)}
-
-			<Modal
-				opened={open}
-				onClose={() => {
-					setOpen(false);
-				}}
-				title={<Title order={2}>Föreslå ny plats</Title>}
-				size="auto"
-			>
-				<BeerLocationForm
-					loading={mutation.isPending}
-					submitButtonText="Skicka"
-					showClearButton={true}
-					onSubmit={onSubmit}
-					districts={districts ?? undefined}
-				/>
-			</Modal>
-		</>
-	);
-};
+export default BeerLocationSubmissionDialog;
