+++ src/routes/__root.tsx
@@ -1,12 +1,8 @@
-import { type Session } from "@supabase/supabase-js";
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
 
 import AppWrapper from "@common/components/app-wrapper";
 import Layout from "@common/components/layout";
-
-interface RouterContext {
-	session: Session | null;
-}
+import { RouterContext } from "@common/types/router";
 
 export const Route = createRootRouteWithContext<RouterContext>()({
 	component: () => (
