+++ src/routes/admin/add-beer-location.tsx
@@ -1,18 +1,7 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 
 import AddBeerLocation from "@feature/admin/pages/add-beer-location";
 
 export const Route = createFileRoute("/admin/add-beer-location")({
 	component: AddBeerLocation,
-	beforeLoad: ({ context, location }) => {
-		if (!context.session) {
-			// eslint-disable-next-line @typescript-eslint/only-throw-error
-			throw redirect({
-				to: "/login",
-				search: {
-					redirect: location.href,
-				},
-			});
-		}
-	},
 });
